Вынес Strategy в отдельный класс и протетсировал.

+++ app/src/main/java/pro/butovanton/weather/strategy/temperatureCalculation.kt
@@ -1,5 +1,5 @@
 package pro.butovanton.weather.strategy
 
 interface temperatureCalculation {
-    fun calculateTemper(tC: Int) : Int
+    fun calculateTemper(tC: Float) : Float
 }
